requirements.txt
anthropic>=0.8.0
pyperclip>=1.8.2
pygame>=2.5.0
pyyaml>=6.0.1
gtts>=2.3.2
keyboard>=0.13.5
pyinstaller>=6.0.0

pywin32==306
keyboard==0.13.5
pyyaml==6.0.1
python-dotenv==1.0.0
anthropic==0.8.1
openai==1.3.7
httpx==0.25.2
aiohttp==3.9.1
asyncio==3.4.3
structlog==24.1.0
colorama==0.4.6
pyinstaller==6.3.0

# UI dependencies
pywebview==4.4.1
jinja2==3.1.3
tailwindcss==3.4.1
pyttsx3>=2.90